requirements.txt
aiogram>=3.0
python-dotenv
SQLAlchemy>=2.0